buildinfo parser: handle version strings from Go dev builds
With the various possible version strings in play, trying to parse
this line is a bit haphazard. We really only need to extract the
executable path, so I've switched to using a regex. That comes at
a cost of less semantic error returns, but as it happens we weren't
receiving these error lines anyway.. they are printed to stderr,
and we only capture stdout from `go version -m ...`. There is the
option to capture stderr as well, but in a way only dealing with
stdout makes life easier as there is less of a chance we mishandle
an error line, etc.

The `verifyExtracted` function already picks up on any paths not
parsed from the output - I've adjusted the error handling slightly so
it's more obvious what has caused the issue.

Resolves #10